# backend/storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fishbook API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/backgrounds": {
            "get": {
                "tags": [
                    "Backgrounds"
                ],
                "operationId": "listBackgrounds",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BackgroundCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/backgrounds/upload": {
            "post": {
                "tags": [
                    "Backgrounds"
                ],
                "operationId": "uploadBackground",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UploadBackgroundRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BackgroundResourceEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/backgrounds/generate": {
            "post": {
                "tags": [
                    "Backgrounds"
                ],
                "operationId": "generateBackground",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenerateBackgroundRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BackgroundResourceEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "429": {
                        "description": "Rate limited"
                    },
                    "503": {
                        "description": "Daily global ceiling exhausted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/backgrounds/{background}/select": {
            "patch": {
                "tags": [
                    "Backgrounds"
                ],
                "operationId": "selectBackground",
                "parameters": [
                    {
                        "name": "background",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid",
                            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BackgroundResourceEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Concurrent select race"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/backgrounds/{background}": {
            "delete": {
                "tags": [
                    "Backgrounds"
                ],
                "operationId": "deleteBackground",
                "parameters": [
                    {
                        "name": "background",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "ulid",
                            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Soft-deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GenerateBackgroundRequest": {
                "required": [
                    "prompt"
                ],
                "properties": {
                    "prompt": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 3
                    },
                    "aspect_ratio": {
                        "type": "string",
                        "default": "16:9",
                        "enum": [
                            "16:9",
                            "3:2",
                            "1:1"
                        ]
                    }
                },
                "type": "object"
            },
            "UploadBackgroundRequest": {
                "required": [
                    "image"
                ],
                "properties": {
                    "image": {
                        "description": "Layered MIME defense: this FormRequest's `mimes:` rule is the cheap\nfirst gate (finfo). The deep gate lives in BackgroundImageProcessor\n(Intervention v3 actually decodes the file). Both must pass.",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "BackgroundResource": {
                "required": [
                    "id",
                    "kind",
                    "storage_key",
                    "signed_url",
                    "width",
                    "height",
                    "is_active",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "ulid",
                        "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$",
                        "example": "01HG5W7DZ8KX9F3N2VYRMPQABT"
                    },
                    "kind": {
                        "type": "string",
                        "enum": [
                            "upload",
                            "generated",
                            "preset"
                        ]
                    },
                    "storage_key": {
                        "type": "string"
                    },
                    "signed_url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "width": {
                        "type": "integer"
                    },
                    "height": {
                        "type": "integer"
                    },
                    "prompt": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "BackgroundCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BackgroundResource"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "BackgroundResourceEnvelope": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/BackgroundResource"
                    }
                },
                "type": "object"
            }
        }
    },
...
}